shift the get repository functions from fetchProject.js to each services files
