+++ services/bitbucketServices.js
@@ -38,6 +38,29 @@ const getTags = async (repoInfo) => {
   }
 };
 
+// function for get project details from bitbucket
+const getRepositoryFromBitbucket = async (project) => {
+  try {
+    const projectRepo = await request.get(
+      `https://api.bitbucket.org/2.0/repositories/${project.handle}/${project.repositorieName}?access_token=${process.env.BITBUCKET_ACCESS_TOKEN}`
+    );
+    if (projectRepo.body) {
+      return projectRepo.body;
+    } else {
+      return false;
+    }
+  } catch (err) {
+    Sentry.captureException(err);
+    logger.error(
+      "Error executing while fetching projects in get repositories from bitbucket function"
+    );
+    logger.error(err);
+    logger.info("=========================================");
+    return false;
+  }
+};
+
 module.exports = {
   getTags: getTags,
+  getRepositoryFromBitbucket: getRepositoryFromBitbucket,
 };
