+++ services/gitlabServices.js
@@ -20,7 +20,40 @@ const getLabels = async (project_id) => {
     return null;
   }
 };
+// function for get project details from gitlab
+const getRepositoryFromGitlab = async (project) => {
+  try {
+    let projectStatus = false;
+    const gitlabUser = await request.get(
+      `https://gitlab.com/api/v4/users?username=${project.handle}`
+    );
+    if (gitlabUser.body.length != 0) {
+      const gitlabRepos = await request
+        .get(
+          `https://gitlab.com/api/v4/users/${gitlabUser.body[0].id}/projects`
+        )
+        .set({ "PRIVATE-TOKEN": process.env.GITLAB_ACCESS_TOKEN });
+      gitlabRepos.body.forEach((repo) => {
+        if (project.repositorieName.localeCompare(repo.path) == 0) {
+          projectStatus = repo;
+        }
+      });
+      return projectStatus;
+    } else {
+      return false;
+    }
+  } catch (err) {
+    Sentry.captureException(err);
+    logger.error(
+      "Error executing while fetching projects in get repositories from gitlab function"
+    );
+    logger.error(err);
+    logger.info("=========================================");
+    return false;
+  }
+};
 
 module.exports = {
   getLabels: getLabels,
+  getRepositoryFromGitlab: getRepositoryFromGitlab,
 };
